ios: clean up
