+++ ios/RTNDatePickerUIViewDelegate.swift
@@ -1,0 +1,7 @@
+import SwiftUI
+
+@objc public protocol RTNDatePickerUIViewDelegate: NSObjectProtocol {
+  @objc func onChange(date: Date)
+  @objc func onConfirm()
+  @objc func onCancel()
+}
