+++ ios/RTNDatePickerView.swift
@@ -1,15 +1,5 @@
 import SwiftUI
 
-class RTNDatePickerViewModel: ObservableObject {
-  @Published var type: String = "date"
-  @Published var isOpen: Bool = false
-  @Published var isInline: Bool = false
-  @Published var value: Date = Date()
-  @Published var range: ClosedRange<Date> = Date.distantPast...Date.distantFuture
-
-  init() {}
-}
-
 struct RTNDatePickerView: View {
   @ObservedObject var viewModel: RTNDatePickerViewModel
   var onChange: (Date) -> Void
